style: Move wishlist entity to wishlist.entity package
